feat: start converting to routify
:D
