+++ .nolluprc
@@ -1,0 +1,7 @@
+{
+  "hot": true,
+  "contentBase": "assets",
+  "publicPath": "build",
+  "historyApiFallback": "__app.html",
+  "port": 5000
+}
