+++ package.json
@@ -2,10 +2,12 @@
   "name": "verto-app",
   "version": "0.0.1",
   "scripts": {
-    "dev": "sapper dev",
-    "build": "sapper build --legacy",
-    "export": "sapper export --legacy --entry \"/ trade tokens post login gallery app 404 app/all-exchanges app/all-transactions\"",
-    "start": "node __sapper__/build",
+    "dev": "run-p routify nollup",
+    "build": "run-s build:*",
+    "build:app": "routify -b && rollup -c",
+    "build:static": "spank",
+    "nollup": "nollup -c",
+    "routify": "routify",
     "prefmt": "cd ./node_modules/prettier-plugin-svelte/ && yarn && yarn build",
     "fmt": "prettier --write .",
     "fmt:check": "prettier --check ."
@@ -40,6 +42,7 @@
     "@rollup/plugin-replace": "^2.2.0",
     "@rollup/plugin-typescript": "^5.0.2",
     "@rollup/plugin-url": "^5.0.1",
+    "@roxi/routify": "^2.7.3",
     "@tsconfig/svelte": "^1.0.10",
     "@types/polka": "^0.5.1",
     "@typescript-eslint/eslint-plugin": "^4.0.1",
@@ -47,14 +50,17 @@
     "eslint": "^7.5.0",
     "eslint-plugin-svelte3": "^2.7.3",
     "node-sass": "^4.14.1",
+    "nollup": "^0.13.13",
+    "npm-run-all": "^4.1.5",
     "prettier": "^2.1.1",
     "prettier-plugin-svelte": "sveltejs/prettier-plugin-svelte#5fbbee9",
     "rollup": "^2.3.4",
+    "rollup-plugin-livereload": "^2.0.0",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-string": "^3.0.0",
     "rollup-plugin-svelte": "^6.0.0",
     "rollup-plugin-terser": "^7.0.0",
-    "sapper": "0.28.7",
+    "spank": "^1.4.0",
     "svelte": "^3.0.0",
     "svelte-check": "^1.0.34",
     "svelte-preprocess": "^4.0.8",
