+++ src/pages/app/all-transactions.svelte
@@ -5,7 +5,7 @@
   import Loading from "../../components/Loading.svelte";
   import SkeletonLoading from "../../components/SkeletonLoading.svelte";
   import moment from "moment";
-  import { goto } from "@sapper/app";
+  import { goto } from "@roxi/routify";
   import { fade } from "svelte/transition";
   import { onMount } from "svelte";
   import { query } from "../../api-client";
@@ -13,7 +13,7 @@
   import allTransactionsQuery from "../../queries/allTransactions.gql";
   import type { Transaction } from "../../utils/types";
 
-  if (process.browser && !$loggedIn) goto("/");
+  if (!$loggedIn) $goto("/");
 
   let client, element;
   let transactions: Transaction[] = [];
@@ -28,8 +28,6 @@
   onMount(() => loadMoreTransactions());
 
   async function loadMoreTransactions() {
-    if (!process.browser) return [];
-
     loading = true;
 
     const client = new Arweave({
