+++ src/pages/gate.svelte
@@ -1,13 +1,10 @@
 <script lang="typescript">
-  import { goto } from "@sapper/app";
+  import { goto } from "@roxi/routify";
 
   import { loggedIn } from "../stores/keyfileStore";
   import Button from "../components/Button.svelte";
 
-  if (process.browser && $loggedIn) {
-    console.log($loggedIn);
-    goto("/app");
-  }
+  if ($loggedIn) $goto("/app");
 </script>
 
 <svelte:head>
@@ -22,7 +19,7 @@
         Verto is in testing and only open to early permaweb adopters right now.
       </p>
       <p>We look forward to welcoming you when you become an Arweaver.</p>
-      <Button click={() => goto('/')}>Go Back</Button>
+      <Button click={() => $goto('/')}>Go Back</Button>
     </div>
   </div>
 </div>
